TPS now reports exact numbers, not to nearest 1000

+++ src/org/voltdb/chargingdemo/BaseChargingDemo.java
@@ -483,7 +483,7 @@ public abstract class BaseChargingDemo {
 
         double tps = tranCount;
         tps = tps / (System.currentTimeMillis() - startMsRun);
-        tps = (long) tps * 1000;
+        tps = tps * 1000;
 
         reportRunLatencyStats(tpMs, tps);
